@@ -1,23 +1,23 @@
 plugins {
-    id "com.jfrog.bintray" version "1.7"
+    id 'signing'
 }
 
 apply plugin: 'com.android.library'
-apply plugin: 'com.github.dcendents.android-maven'
 
-group 'com.transloadit.android.sdk'
-version = '0.0.4'
-description = "An Android Integration of the Transloadit's(https://transloadit.com) file uploading and encoding service."
+repositories {
+    mavenCentral()
+    jcenter()
+}
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "26.0.2"
+    compileSdkVersion 29
+    buildToolsVersion "30.0.2"
     defaultConfig {
         minSdkVersion 15
-        targetSdkVersion 24
+        targetSdkVersion 29
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
