@@ -1,14 +1,20 @@
 package com.transloadit.android.sdk;
 
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+
 import android.app.Activity;
+
 
 import com.transloadit.sdk.async.AssemblyProgressListener;
 import com.transloadit.sdk.response.AssemblyResponse;
 
+import org.junit.After;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.Mockito;
-import org.mockserver.client.server.MockServerClient;
+import org.mockserver.client.MockServerClient;
 import org.mockserver.junit.MockServerRule;
 import org.mockserver.model.HttpRequest;
 import org.mockserver.model.HttpResponse;
@@ -17,17 +23,13 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
 
 import io.tus.java.client.TusURLMemoryStore;
 
-import static org.junit.Assert.*;
-import static org.mockserver.model.RegexBody.regex;
-
 public class AndroidAsyncAssemblyTest {
+    public final int PORT = 9040;
     @Rule
-    public MockServerRule mockServerRule = new MockServerRule(9040, this, true);
+    public MockServerRule mockServerRule = new MockServerRule(this, true, PORT);
 
     private MockServerClient mockServerClient;
     private boolean uploadFinished;
@@ -36,14 +38,13 @@ public class AndroidAsyncAssemblyTest {
     private Exception statusUpdateError;
     private Exception uploadError;
 
+
     @Test
     public void testSave() throws Exception {
         // for assembly creation
         mockServerClient.when(HttpRequest.request()
-                .withPath("/assemblies")
-                .withMethod("POST")
-                .withBody(regex("[\\w\\W]*tus_num_expected_upload_files\"\\r\\nContent-Length: 1" +
-                        "\\r\\n\\r\\n1[\\w\\W]*")))
+                .withPath("/assemblies/76fe5df1c93a0a530f3e583805cf98b4")
+                .withMethod("POST"))
                 .respond(HttpResponse.response().withBody(getJson("assembly.json")));
 
         // for assembly status check
@@ -51,13 +52,13 @@ public class AndroidAsyncAssemblyTest {
                 .withPath("/assemblies/76fe5df1c93a0a530f3e583805cf98b4").withMethod("GET"))
                 .respond(HttpResponse.response().withBody(getJson("assembly.json")));
 
-        AndroidTransloadit transloadit = new AndroidTransloadit("KEY", "SECRET", "http://localhost:9040");
+        AndroidTransloadit transloadit = new AndroidTransloadit("KEY", "SECRET", "http://localhost:" + PORT);
         AssemblyProgressListener listener = new Listener();
         AndroidAsyncAssembly assembly = new MockAsyncAssembly(transloadit, listener, Mockito.mock(Activity.class));
+        assembly.setAssemblyId("76fe5df1c93a0a530f3e583805cf98b4");
         assembly.setTusURLStore(new TusURLMemoryStore());
         assembly.addFile(new File(getClass().getClassLoader().getResource("assembly.json").getFile()), "file_name");
         AssemblyResponse resumableAssembly = assembly.save();
-
         assertEquals(resumableAssembly.json().get("assembly_id"), "76fe5df1c93a0a530f3e583805cf98b4");
         assertTrue(uploadFinished);
         assertTrue(assemblyFinished);
@@ -73,7 +74,7 @@ public class AndroidAsyncAssemblyTest {
         }
 
         @Override
-        public void onUploadPogress(long uploadedBytes, long totalBytes) {
+        public void onUploadProgress(long uploadedBytes, long totalBytes) {
             totalUploaded = uploadedBytes;
         }
 
